Update client.lua
Added Client Event for updateAddMoney to allow the + icon to be displayed when someone uses the export and TriggerClientEvent from a script in which the player would receive money. Currently the only option was to use TriggerClientEvent("updateMoney" xxxx) which only allows for a - symbol even when money is being added. Now they can use TriggerClientEvent("updateAddMoney" xxxx) when the player should receive money so it will show the + symbol. (edit to server.lua to follow immediately after this)